change view but broke the app functionality
