+++ src/server/routes.go
@@ -1,12 +1,11 @@
 package server
 
 import (
+	"custom-db-platform/src/server/handlers"
 	"custom-db-platform/src/web"
 	"fmt"
-	"html/template"
 	"log"
 	"net/http"
-	"sync"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
@@ -17,18 +16,6 @@ type userDetails struct {
 	LastName  string
 }
 
-type Result struct {
-	Message string
-	Success bool
-}
-
-type filteredResults struct {
-	Sucesses []string
-	Errors   []string
-}
-
-var wg sync.WaitGroup
-
 func (s *Server) RegisterRoutes() http.Handler {
 
 	router := chi.NewRouter()
@@ -37,9 +24,9 @@ func (s *Server) RegisterRoutes() http.Handler {
 	router.Group(func(r chi.Router) {
 		// r.Use(utils.CustomMiddleware())
 
-		r.Get("/", homeHandler)
+		r.Get("/", handlers.LoadHomePage)
 
-		r.Get("/create-user", createUserPageHandler)
+		r.Get("/create-user", handlers.LoadCreateUserForm)
 		r.Post("/create-user", createUserFormHandler)
 
 		r.Get("/delete-user", deleteUserPageHandler)
@@ -47,79 +34,57 @@ func (s *Server) RegisterRoutes() http.Handler {
 
 		r.Get("/configuration", configPageHandler)
 		r.Post("/db", addDBHandler)
+
+		r.Get("/sign-in", handleSignIn)
 	})
 
 	return router
 }
 
-func homeHandler(w http.ResponseWriter, r *http.Request) {
+// func createUserFormHandler(w http.ResponseWriter, r *http.Request) {
+// 	r.ParseForm()
+// 	username := r.FormValue("username")
+// 	wo := r.FormValue("wo")
+// 	databases := r.Form["databases"]
+// 	fmt.Println(databases)
 
-	// TODO: Do I need the user details? If yes where am I going to get it from?
-	// the request context?
-	// ctx := r.Context()
-	// sessClaims, _ := ctx.Value(clerk.ActiveSessionClaims).(*clerk.SessionClaims)
+// 	var results []Result
 
-	// user, err := clerkClient.Users().Read(sessClaims.Subject)
-	// if err != nil {
-	// 	panic(err)
-	// }
+// 	c := make(chan Result)
 
-	web.Templates["home"].ExecuteTemplate(w, "base-layout.tmpl", nil)
-}
-func createUserPageHandler(w http.ResponseWriter, r *http.Request) {
+// 	for _, database := range databases {
+// 		wg.Add(1)
 
-	dbs, err := getDBsName()
-	if err != nil {
-		log.Fatal(err)
-	}
+// 		dbDetail, err := getDBByName(database)
 
-	web.Templates["createUser"].ExecuteTemplate(w, "base-layout.tmpl", dbs)
-}
+// 		if err != nil {
+// 			log.Fatal(err)
+// 		}
 
-func createUserFormHandler(w http.ResponseWriter, r *http.Request) {
-	r.ParseForm()
-	username := r.FormValue("username")
-	wo := r.FormValue("wo")
-	databases := r.Form["databases"]
-	fmt.Println(databases)
+// 		fmt.Printf("username: %s, wo: %s, database: %v\n", username, wo, dbDetail)
+// 		go ConnectToDBAndCreateUser(dbDetail.Host, dbDetail.Port, dbDetail.DbType, dbDetail.SslMode, username, dbDetail.Name, c)
+// 		msg := <-c
+// 		results = append(results, msg)
+// 	}
+// 	wg.Wait()
 
-	var results []Result
+// 	var fResponse filteredResults
 
-	c := make(chan Result)
+// 	for _, result := range results {
+// 		if result.Success {
+// 			fResponse.Sucesses = append(fResponse.Sucesses, result.Message)
+// 		} else {
+// 			fResponse.Errors = append(fResponse.Errors, result.Message)
+// 		}
+// 	}
 
-	for _, database := range databases {
-		wg.Add(1)
+// 	tmpl, err := template.ParseFiles("src/web/response.tmpl")
 
-		dbDetail, err := getDBByName(database)
-
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		fmt.Printf("username: %s, wo: %s, database: %v\n", username, wo, dbDetail)
-		go ConnectToDBAndCreateUser(dbDetail.Host, dbDetail.Port, dbDetail.DbType, dbDetail.SslMode, username, dbDetail.Name, c)
-		msg := <-c
-		results = append(results, msg)
-	}
-	wg.Wait()
-
-	var fResponse filteredResults
-
-	for _, result := range results {
-		if result.Success {
-			fResponse.Sucesses = append(fResponse.Sucesses, result.Message)
-		} else {
-			fResponse.Errors = append(fResponse.Errors, result.Message)
-		}
-	}
-
-	tmpl, err := template.ParseFiles("src/web/response.tmpl")
-
-	if err != nil {
-		log.Fatal(err)
-	}
-	err = tmpl.Execute(w, fResponse)
-}
+// 	if err != nil {
+// 		log.Fatal(err)
+// 	}
+// 	err = tmpl.Execute(w, fResponse)
+// }
 
 func deleteUserPageHandler(w http.ResponseWriter, r *http.Request) {
 
@@ -160,4 +125,8 @@ func addDBHandler(w http.ResponseWriter, r *http.Request) {
 	}
 	w.WriteHeader(http.StatusCreated)
 	w.Write([]byte(fmt.Sprintf("<div class=\"border border-green-500 bg-green-300 w-fit p-2 rounded\">%s has been created successfully.</div>", name)))
+}
+
+func handleSignIn(w http.ResponseWriter, r *http.Request) {
+	web.Templates["signIn"].ExecuteTemplate(w, "base-layout.tmpl", nil)
 }
