+++ src/web/templates/dbConfig/addDbForm.tmpl
