+++ src/web/templates/dbUser/deleteDbUserForm.tmpl
